+++ tests/BaseConvertTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Phlib;
 
+use PHPUnit\Framework\Error\Deprecated as DeprecatedError;
 use PHPUnit\Framework\TestCase;
 
 class BaseConvertTest extends TestCase
@@ -12,6 +13,57 @@ class BaseConvertTest extends TestCase
     {
         $number = '1234567890';
         static::assertSame(\base_convert($number, 10, 36), \Phlib\base_convert($number, 10, 36));
+    }
+
+    /**
+     * @dataProvider dataValidNumberCharacters
+     */
+    public function testValidNumberCharacters(string $number, bool $isValid): void
+    {
+        if (!$isValid) {
+            $this->expectException(\InvalidArgumentException::class);
+            $this->expectExceptionMessage('Invalid characters passed for attempted conversion');
+        }
+
+        \Phlib\base_convert($number, 36, 10);
+
+        if ($isValid) {
+            // Valid number should reach this line without error
+            static::assertTrue(true);
+        }
+    }
+
+    public function dataValidNumberCharacters(): array
+    {
+        return [
+            'numeric' => ['123', true],
+            'alphanumeric' => ['123abc', true],
+            'float' => ['3.6', false],
+            'scientific' => ['1.2e3', false],
+        ];
+    }
+
+    /**
+     * @requires PHP 7.4
+     */
+    public function testOriginalNumberCharactersWarning(): void
+    {
+        /**
+         * PHP's own validation for `$number` is more accurate
+         * as it warns against characters that are not allowed by the given from base,
+         * e.g. *base2* only allows *0* and *1*.
+         */
+        $originalErrorLevel = error_reporting();
+        error_reporting(E_ALL);
+
+        $this->expectException(DeprecatedError::class);
+        $this->expectExceptionMessage('Invalid characters passed for attempted conversion, these have been ignored');
+
+        try {
+            \Phlib\base_convert('123', 2, 10);
+        } finally {
+            error_reporting($originalErrorLevel);
+        }
     }
 
     /**