+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Type declarations have been added to all parameters and return types.
 - Validation with `\InvalidArgumentException` for the `base_convert()`
   parameters to match the warnings added for the original PHP function in v7.4.
+  - `$number` must be a string using only *alphanumeric* characters.
   - `$fromBase` and `$toBase` must be an integer between *2* and *36*.
 ### Removed
 - **BC break**: Removed support for PHP versions < v7.1 as they are no longer